AM-73 : The command line "list --no-compat" cannot be executed
AM_LIST_10 Unless --no-compatoption is passed, the list is filtered to match the app server  with supportedApplicationServers.
Example : if supportedApplicationServers contains tomcat and the app server is JBoss, the entry will not be listed

AM_LIST_11 Unless --no-compatoption is passed, the list is filtered to match the eXo Platform edition with supportedDistributions.
Example : if supportedDistributions contains enterprise, the entry will not be listed on a eXo Platform Community Edition.

AM_LIST_12 Unless --no-compatoption is passed, the list is filtered to match the eXo Platform version with compatibility.
Example : if compatibility contains [4.1-M2,), the entry will not be listed on a eXo Platform 4.0