AM-74 : Cannot install an addon while ignoring compatibility check
AM_INST_09  :
The add-ons manager does a compatibility check using the compatibility, applicationServer and distributions values prior to install an add-on. If the add-on is not compatible, the installation interrupts with an error of the form :
The add-on $ID is not compatible : $REASON. Use --no-compatto bypass this compatibility check and install anyway. where

$ID is the addonId:version
$REASON is the reason of the compatibility check (e.g : app server jboss not supported, distribution enterprise required, not compatible with version 4.0 )